Separate ZAPI into their own files
